+++ src/main/java/naem/server/repository/custom/CustomPostRepository.java
@@ -1,4 +1,4 @@
-package naem.server.repository;
+package naem.server.repository.custom;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
