+++ src/main/java/naem/server/web/BoardController.java
@@ -62,8 +62,8 @@ public class BoardController {
 
     @ApiOperation(value = "게시글 단건 조회", notes = "게시글 단건 조회")
     @GetMapping("/detail/{id}")
-    public DetailedPostInfoDto getPost(@PathVariable("id") long id) {
-        return postService.getDetailedPostInfo(id);
+    public DetailedPostInfoDto getPost(@PathVariable("id") long postId) {
+        return postService.getDetailedPostInfo(postId);
     }
 
     @ApiOperation(value = "게시글 수정", notes = "게시글 수정")
